+++ javax/src/main/java/me/panpf/javax/util/Predicate.java
@@ -19,5 +19,5 @@ package me.panpf.javax.util;
 import org.jetbrains.annotations.NotNull;
 
 public interface Predicate<T> {
-    boolean predicate(@NotNull T t);
+    boolean accept(@NotNull T t);
 }