+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Kotlin 版本移除 annotations 依赖
 * Comparisonx 增加 maxOf, minOf
 * Arrayx 增加 zip
+* Arrayx 增加 partition
 
 ## 0.6.1
 * Arrayx 增加 plus, contentToString